feat(docs): reorganize documentation structure and fix build errors
Documentation reorganization:
- Move implementation plans to docs/plans/
  * IMPLEMENTATION_PLAN_RPI5_HARDWARE.md
  * IMPLEMENTATION_PLAN_SISLANG.md
  * IMPLEMENTATION_PLAN_STUBS_TO_REAL.md
  * IMPLEMENTATION_PLAN_X86_64.md
- Create docs/implementation/rpi5/ directory
- Move all RPI5 documentation to docs/implementation/rpi5/
  * RPI5_HARDWARE_IMPLEMENTATION.md
  * RPI5_IMPLEMENTATION_SUMMARY.md
  * RPI5_INTEGRATION_GUIDE.md
  * RPI5_M1_STORAGE.md
  * RPI5_M2_POWER.md
  * RPI5_M3_SMP.md
  * RPI5_M7_VALIDATION.md
  * RPI5_M8_HARDENING.md
  * RPI5_M8_HARDENING_SUMMARY.md
  * RPI5_REVIEW_M0_M1.md

Build fixes:
- Add parse_number() helper to shell.rs for hex/dec parsing
- Fix compilation errors across RPI5 driver modules
- Update imports and error handling in platform/validation code

README updates:
- Update phase status descriptions from 'PRODUCTION READY' to
  'Implemented and exercised in QEMU' for accuracy

+++ crates/kernel/src/pmu.rs
@@ -44,6 +44,7 @@ static PMU_INITIALIZED: AtomicBool = AtomicBool::new(false);
 #[cfg(target_arch = "aarch64")]
 pub mod aarch64 {
     use super::PMU_INITIALIZED;
+    use super::{DriverError, DriverResult, Validator};
     use core::sync::atomic::Ordering;
     #[inline(always)]
     unsafe fn write_pmselelr(idx: u64) {
@@ -320,7 +321,7 @@ pub mod aarch64 {
         }
 
         // M8: Validate counter index (0-5)
-        Validator::check_bounds(counter_idx, 0, super::MAX_EVENT_COUNTER)?;
+        Validator::check_bounds(counter_idx as usize, 0, super::MAX_EVENT_COUNTER as usize)?;
 
         write_pmselelr(counter_idx);
         Ok(read_pmxevcntr())
@@ -330,7 +331,7 @@ pub mod aarch64 {
 #[cfg(not(target_arch = "aarch64"))]
 pub mod aarch64 {
     use super::PMU_INITIALIZED;
-    use crate::drivers::{DriverError, DriverResult, Validator};
+    use super::{DriverError, DriverResult, Validator};
     use core::sync::atomic::Ordering;
 
     #[derive(Copy, Clone, Default, Debug)]
@@ -431,4 +432,3 @@ pub fn read_event_counter(counter_idx: u64) -> DriverResult<u64> {
 pub fn is_initialized() -> bool {
     PMU_INITIALIZED.load(Ordering::Acquire)
 }
-
